Feature/hist seasons (#29)
* hist seasons

* hist sourcing test run

* prevent downstream tiers from processing

* stop lower tier processing

* missed splitQuery

* leaf addition

* extract classes to new files

* historical season sourcing

* pr-based changes

* pr-based change - race con

* chg resourceIndex id datatype

+++ src/SportsData.Provider/Infrastructure/Data/Entities/ResourceIndex.cs
@@ -10,7 +10,7 @@ namespace SportsData.Provider.Infrastructure.Data.Entities
 {
     public class ResourceIndex : CanonicalEntityBase<Guid>, IHasSourceUrlHash
     {
-        public int Ordinal { get; set; }
+        public long Ordinal { get; set; }
 
         public required string Name { get; set; }
